Abstracted duplicated code to AbstractGitHubService
